+++ source/components/player/comp_player_model.cpp
@@ -108,7 +108,7 @@ void TCompPlayerModel::onLevelChange(const TMsgPowerLvlChange& msg) {
 	currentPowerStats = powerStats[msg.powerLvl - 1];
 
 	TCompRender *render = get<TCompRender>();
-	render->setAllMaterials(0, render->meshes.size() / 2, materials[msg.powerLvl - 1]);
+	render->setAllMaterials(0, (int)render->meshes.size() / 2, materials[msg.powerLvl - 1]);
 
 	EngineScripting.throwEvent(onPowerLevelChange, std::to_string(msg.powerLvl));
 }
@@ -297,7 +297,7 @@ void TCompPlayerModel::setRespawnPosition(VEC3 position) {
 
 void TCompPlayerModel::disableOutline() {
 	TCompRender *render = get<TCompRender>();
-	for (int i = render->meshes.size() / 2; i < render->meshes.size(); ++i) {
+	for (size_t i = render->meshes.size() / 2; i < render->meshes.size(); ++i) {
 		render->meshes[i].enabled = false;
 	}
 	render->refreshMeshesInRenderManager();
@@ -305,7 +305,7 @@ void TCompPlayerModel::disableOutline() {
 
 void TCompPlayerModel::enableOutline() {
 	TCompRender *render = get<TCompRender>();
-	for (int i = render->meshes.size() / 2; i < render->meshes.size(); ++i) {
+	for (size_t i = render->meshes.size() / 2; i < render->meshes.size(); ++i) {
 		render->meshes[i].enabled = true;
 	}
 	render->refreshMeshesInRenderManager();
